Add a Capacity Provider resource to the Auto Scaling Group because that's how AWS does this now.
This created a dependency loop because the instance userdata (and
by extension the Capacity Provider) depends on the name of the
ECS Cluster. The ECS Cluster in turn depends on the name of the
Capacity Provider. I broke the loop by putting the desired cluster
name into a local variable and then used the local variable to
define the instance userdata and the ECS Cluster.

I added protection from scale in and the AmazonECSManaged tag to
the Auto Scaling Group because that's what's needed to make the
Capacity Provider work properly.

Added a new variable called container_insights that can be set
to "enabled" or "disabled". The default is "enabled".